+++ kernel/arch/amd64/acpi/acpi.h
