+++ kernel/arch/cpu.h
@@ -23,6 +23,7 @@
  * arch_idle() - idle CPU (enable interrupts and halt)
  * arch_pause() - hint to CPU that we're in a spin loop
  * arch_set_kernel_stack(void *stack_top) - set kernel stack for ring transitions
+ * arch_cpu_index() - get the current CPU logical index/ID
  *
  * memory barriers:
  *
